Update the BSON implementation to use the new generic binary type
BSON used to use 0x02 as the "binary subtype" identifier for generic
binary strings. Since then, it has been updated to use 0x00, so we
should use that by default. Reading BSON strings using the old format
should still work, so this just converts them to the new format
automatically. However, it is now impossible to write strings in the old
format.